Fundamentos pt3 - Aula 05
Refatoração e delegação de responsabilidades
