+++ Agenda/app/src/main/java/br/com/alura/agenda/ui/ListaAlunosView.java
@@ -1,0 +1,59 @@
+package br.com.alura.agenda.ui;
+
+import android.app.AlertDialog;
+import android.content.Context;
+import android.content.DialogInterface;
+import android.view.MenuItem;
+import android.widget.AdapterView;
+import android.widget.ListView;
+
+import androidx.annotation.NonNull;
+
+import br.com.alura.agenda.dao.AlunoDAO;
+import br.com.alura.agenda.model.Aluno;
+import br.com.alura.agenda.ui.adapter.ListaAlunosAdapter;
+
+public class ListaAlunosView {
+
+    private final ListaAlunosAdapter adapter;
+    private final AlunoDAO dao;
+    private final Context context;
+
+    public ListaAlunosView(Context context) {
+        this.context = context;
+        this.adapter = new ListaAlunosAdapter(this.context);
+        this.dao = new AlunoDAO();
+    }
+
+    public void confirmaRemocao(@NonNull final MenuItem item) {
+        new AlertDialog
+                .Builder(context)
+                .setTitle("Removendo aluno")
+                .setMessage("Tem certeza que deseja remover o aluno?")
+                .setPositiveButton("Sim", new DialogInterface.OnClickListener() {
+                    @Override
+                    public void onClick(DialogInterface dialog, int which) {
+                        AdapterView.AdapterContextMenuInfo menuInfo = (AdapterView.AdapterContextMenuInfo) item.getMenuInfo();
+
+                        Aluno aluno = adapter.getItem(menuInfo.position);
+                        remove(aluno);
+                    }
+                })
+                .setNegativeButton("Não",null)
+                .show();
+    }
+
+    public void atualizaAlunos() {
+        adapter.atualiza(dao.todos());
+    }
+
+    private void remove(Aluno aluno) {
+        dao.remove(aluno);
+        adapter.remove(aluno);
+    }
+
+    public void configuraAdapter(ListView listaAlunos) {
+        listaAlunos.setAdapter(adapter);
+    }
+
+}
