+++ Agenda/app/src/main/java/br/com/alura/agenda/ui/activity/ConstantesActivies.java
@@ -1,5 +1,5 @@
 package br.com.alura.agenda.ui.activity;
 
-public interface ConstantesActivies {
+interface ConstantesActivies {
     String CHAVE_ALUNO = "aluno";
 }
